Moving fetchFromSource to a new directory
